feat: rename files in `hmr` to match kebab-case
