+++ packages/hmr/lib/injections/refresh.ts
@@ -1,4 +1,4 @@
-import initClient from '../initializers/initClient.js';
+import initClient from '../initializers/init-client.js';
 
 (() => {
   let pendingReload = false;
